usb: usb_stress_io: Reload necessary modules with appropriate parameters
Some tests like ISO need specific parameters for the usbtest module so
we reload the modules with the necessary parameters.

Check if USB_TEST and USB_ZERO drivers are configured as modules since we
need to load to load them with different parameters for different test cases.

Increase USB_L_STRESS_UNLINK timeout to 200 seconds.

get rid of XHCI* test cases as it adds no differentiation at the moment.

Signed-off-by: Roger Quadros <rogerq@ti.com>
Signed-off-by: Yan Liu <yan-liu@ti.com>