common: functions.sh: Restart rtcwake on late int
It is possible that a late interrupt will cause the suspend path to
abort. In such scenario, the current suspend code will fail the test.

This patch detects such condition and it starts a new rtcwake call.
It is believe that one retry such be enough to handle this late
interrupts scenario.

Signed-off-by: Carlos Hernandez <ceh@ti.com>