Add new text parsing functions
This patch adds new functions to parse text into lists based on
patterns that can be used to differentiate between sections in
the text. These functions allows the user to create the list,
get the text associated with a section and get a list
of the keys in the list that index each section.

Signed-off-by: Alejandro Hernandez <ajhernandez@ti.com>
Signed-off-by: Yan Liu <yan-liu@ti.com>