usb: usb_perf_raw: Make ISO test usable on XHCI host at High Speed
In ISO mode (i.e. g_zero gadget in Source/Sink configuration with
alternate setting 1), we have 4 endpoints:
1 BULK IN, 1 BULK OUT, 1 ISO IN and 1 ISO OUT.

If we allocate 3072 bytes per micro-frame (maxpacket:1024, mult:2[+1])
for both ISO endpoints then we have 82% of bandwidth of a micro-frame
reserved only for ISO (periodic) endpoints. Estimated from table [1].

As per the USB 2.0 specifications [2], we must not allocate more than
80% of micro-frame bandwidth for periodic endpoints.

This is the reason the XHCI host's bandwidth allocation fails with
"Not enough bandwidth for device state"

To fix this issue we reduce the maxpacket size from 1024 to 960 and so
use only 76.8% of micro-frame bandwidth for periodic endpoints.

[1] USB 2.0 spec - Table 5-5. High-speed Isochronous Transaction Limits
[1] USB 2.0 spec - 5.6.4 Isochronous Transfer Bus Access Constraints

Signed-off-by: Roger Quadros <rogerq@ti.com>
Signed-off-by: Carlos Hernandez <ceh@ti.com>