alsa: Dynamically determine the audio device
This patch adds an option to pass the audio device when testing
alsa volume, switch toggle and performance. It also determines the
default audio device dynamically in case the option is not passed.
Before this patch the default hw:0,0 device was used, which may not
worked for platforms with multiple playout devices like dra7xx-evm

Signed-off-by: Alejandro Hernandez <ajhernandez@ti.com>
Signed-off-by: Yan Liu <yan-liu@ti.com>