Added create two partitions function to create two partition in storage media.
Currently, for emmc, we only create one partition; but the problem was that env variable
use some initial space. When saveenv in uboot, this partition got corrupted.
This patch is to solve this problem by creating two partitions; the first one
is marked as boot partition so the test script won't touch it. The second one is
used for testing.

Signed-off-by: Yan Liu <yan-liu@ti.com>