ipc: Add rpmsg sample client module to test suite
This patch adds the rpmsg sample client kernel module to the ipc test
suite. This module is part of the samples provided by the linux kernel
for rpmsg and allows for basic validation of the rpmsg implementation.
The patch also rearranges the existing ipc kernel modules so that it
is easier to remove/add ipc modules to the ltp-ddt build

Signed-off-by: Alejandro Hernandez <ajhernandez@ti.com>
Signed-off-by: Yan Liu <yan-liu@ti.com>