runltp: Use mktemp to create LOGFILE
Use mktemp instead of using timestamp and process ID to create a
unique logfile. On embedded systems without batteries or NTP support,
the date might be restarted to the same time when the board is
rebooted hence timestamps are not guaranteed to be unique. Similarly,
PIDs are not guaranteed to be unique either. Using mktemp has the
extra benefit that logs will be automatically purged on boot if /tmp
is mounted to RAM as it usually is.

Signed-off-by: Carlos Hernandez <ceh@ti.com>
Signed-off-by: Yan Liu <yan-liu@ti.com>