Fix to take care of corner case of cpu affinity
Per definition of cpu affinity in manual page of taskset:
a cpu affinity of 0x00000003 stands for "both processors #0 and #1"
a cpu affinity of 0xFFFFFFFF stands for "all processors (#0 through #31)"

In a dual core system, setting a cpu affinity of 0xFFFFFFFF using taskset
results in taskset returning a cpu affinity of 3.

Also, taskset of 0xFFFFFFFF actually returns an affinity of F in systems
with more than two cores, so the expected affinity is set accordingly.
This output affinity of f being observed in quad core embedded systems
as well as quad core PCs.
Prior to conversion to decimal value, a '0x' is also appended to
reported affinity since taskset does not print with 0x prefix.
This prefix is required for proper conversion of hex numbers like 'f' to
decimal using printf directive.

This patch takes care of these corner cases so that the test does not result
in a false fail.

Reference: manpages.ubuntu.com/manpages/natty/man1/taskset.1.html

Signed-off-by: Aparna Balasubramanian <aparnab@ti.com>
Signed-off-by: Yan Liu <yan-liu@ti.com>