common/functions.sh: Do not always kill memtest
Currently the EXIT signal is trapped in global context and its
handler kills memtest and cpuloadgen processes. This has the side
effect of unexpectly killing memtest for cases where functions.sh
is sourced to use other non-memtest-related functions.
This patch only register the exit handler when memtest or cpuloadgen
functions are called.

Signed-off-by: Carlos Hernandez <ceh@ti.com>