Test cases for simultaneous dual mac ethernet throughput measurement
Adds ltp tests for tcp and udp simultaneous dual mac ethernet throughput
measurement. Includes supporting test script for determining number of
ethernet interfaces on DUT as well as the names of those interfaces.

These tests can be used as examples to create test cases for n number of
simultaneous ethernet interface throughput tests.

These tests need to be used in combination with corresponding changes to
vatf-scripts application which will be submitted as a separate patch.

Signed-off-by: Aparna Balasubramanian <aparnab@ti.com>
Signed-off-by: Yan Liu <yan-liu@ti.com>