blk_device_common.sh: Add function to create three partitions in storage devices like usbhost, emmc, msata/sata etc.
Currently, usbhost, sata etc tests require the user manually create some partitions in
storage devices in order to run ltp-ddt test. The test would fail if there is no partition.
This patch adds function to automatically create three partitions if there is no
partitions in storage device. This function takes the 1st and 2nd partition size as input,
the 3rd partition will take the remaining. This function also mark the first partition to be
boot partition in case user like to use it later on. The test script only choose the partition
which is not boot partition or rootfs partition to use for testing.

Signed-off-by: Yan Liu <yan-liu@ti.com>