get_blk_device_node.sh: Dynamically get blk device node used for testing instead of using partition 1
Currently, ltp-ddt always use partition 1 to test usbhost and msata/sata.
This patch is to fix this by reusing the same logic as used for mmc. The script
choose the biggest partition which are not boot partition or rootfs partition.
Also, the script will create three partitions if there is no partition in usbhost or sata device.

Signed-off-by: Yan Liu <yan-liu@ti.com>