Accommodate phy related delays in ethernet testing
Certain ethernet phy exhibit delay in coming up, causing
some of the interface up/down tests to fail. The changes
here take into account such cases and modify the
commands sent to dut for putting interface up and down
for such phy types.
Tested on a board with a phy requiring the new sequence and
a board which does not require it.

Signed-off-by: Aparna Balasubramanian <aparnab@ti.com>
Signed-off-by: Yan Liu <yan-liu@ti.com>