gst: Refactor fps calculation for decode tests
This patch refactors the gst fps calculation used in decode
test cases. Before this patch the fps calculation was
performed based on the values returned by fpsdisplaysink
unfortunately this component is not available on all
platforms and returns 0 fps values for some streams. This
patch refactors the test scripts by removing the -p option
which sets the fpsdisplaysink component and refactoring the
fps calculation based on the pipeline execution time

Signed-off-by: Alejandro Hernandez <ajhernandez@ti.com>
Signed-off-by: Carlos Hernandez <ceh@ti.com>