gst_dec: Disable audio in performance test
This patch modifies the gstreamer decode performance test
so that it disables audio playout by specifying fakesink
as the audio playout sink. This is needed because audio
sinks in gstreamer do not provide the option to run without
synchronization. The patch also replaces _PERF with
_VIDEO_PERF in the test tags to indicate that only video
performance is being validated in these cases

Signed-off-by: Alejandro Hernandez <ajhernandez@ti.com>
Signed-off-by: Carlos Hernandez <ceh@ti.com>