ipc: Make sure mpmsrv is terminated
This patch adds an additional killall command to make sure
that the mpmsrv daemon is terminated before removing the
remoteproc modules. Before this patch the test was only
relying on mpm init.d script to terminate the daemon which
did not always worked. This patch are replaces rmmod calls
with modprobe -r which fails to remove the modules if mpmsrv
is still running

Signed-off-by: Alejandro Hernandez <ajhernandez@ti.com>
Signed-off-by: Yan Liu <yan-liu@ti.com>