Provide ability to run ethernet test by type of module
Adds optional parameter 'm' to eth_multi.sh which is used to select an
ethernet interface by module type. For instance, cpsw or pru can
be inputs for -m parameter and when such a parameter is provided,
test will be run only on ethernet interfaces of that type.
This will help run tests separately on cpsw and icss pru interfaces
on the same board, if so preferred.

Signed-off-by: Aparna Balasubramanian <aparnab@ti.com>
Signed-off-by: Yan Liu <yan-liu@ti.com>