video: Update ps call on gst scripts
This patch adds options -ef to the ps call made in the
gst scripts. The change is required after an update
to ps was made. Without these options, on the new version
of ps, the output of the call is very limited and weston
is not stop which causes the gstreamer pipelines to fail

Signed-off-by: Alejandro Hernandez <ajhernandez@ti.com>
Signed-off-by: Yan Liu <yan-liu@ti.com>