Supportive ethernet test scripts for PRU test cases
PRU+CPSW scenarios could result in a pru and a cpsw in the same subnet
on dut board.In such cases, determining gateway of interface under test
requires default gateway to be deleted. Add a simple script to do that.
Some details below.

The test setup allows all available ports on the board to be connected
to two subnets and removes the requirement of having to use a unique subnet
for each cpsw/pru port or having only two ports connected at a time.
So, for a 2 cpsw+2 pru setup, one set of cpsw and pru can be on one subnet
and the other pair can be on second subnet.
Test involves determining gateway for each port. Then, bringing down
all interfaces except the one under test. Run ping tests to gateway
corresponding to interface under test with different duration or packet size
parameters.
For determining gateway, details below.
Routing table does not add a gateway entry if one exists for the same
subnet from a previous interface. Even bringing down the previous
interface does not remove this entry. Hence, to accurately determine
the gateway for a given interface, it is best to remove any default
entry and let the bring up of the interface populate the default gateway.
So, the above will let us run pru tests on setups with cpsw also
connected on multiple ports, thus eventually let it run on a farm
setup with appropriate network connections.

Also, while running PRU scenarios, get_eth_iface_by_module_name.sh can be used
to check that the interface is a pru interface. So, add the script.

Signed-off-by: Aparna Balasubramanian <aparnab@ti.com>
Signed-off-by: Carlos Hernandez <ceh@ti.com>