syscalls/mount03: close files in MS_NOATIME test error paths
The mount03 test has a bug in the error handling code paths for the
MS_NOATIME flag test: if for instance a read() fails then we return
from test_rwflag() without doing a close() on the filedescriptor.
This then causes the umount() performed by tst_release_device() to
fail with EBUSY, and then the loopback device is left mounted.
Later, other test cases that try to use the loopback device then
fail unnecessarily.

Close open file descriptors in error-exit codepaths, so that cleanup
on failure works as intended.  This brings the code for MS_NOATIME
into line with how the other test cases in the function handle this.

Signed-off-by: Peter Maydell <peter.maydell@linaro.org>
Acked-by: Cyril Hrubis <chrubis@suse.cz>