network/tcp_cmds/arp01: use test_net.sh and add test-cases with 'ip neigh'
arp01 stress test removes/adds arp entries on the management interface.
This is unwelcome. Moreover, arp command is obsolete and there is
'ip neigh' command that replaces it.

The patch rewrites the test so that it can test both commands in the
same way + test IPv6 ndisc cache entries. Additionally, it checks that
arp/ndisc entry was removed by searching for rhost's hwaddr in the cache.

Signed-off-by: Alexey Kodanev <alexey.kodanev@oracle.com>
Reviewed-by: Cyril Hrubis <chrubis@suse.cz>