ipc/ipc_stress: Remove.
It has been discussed on GitHub and on ML:

https://github.com/linux-test-project/ltp/issues/75
http://lists.linux.it/pipermail/ltp/2016-July/002209.html
http://lists.linux.it/pipermail/ltp/2016-August/002509.html

In short, these testcases were not really ported to LTP, are not
designed to be executed separately (as it's done in the runtest script),
does not use TMPDIR, etc.

Also the same functionality is covered by corresponding syscall tests,
so there is no need to keep these around.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>