cpuacct_task: exit cleanly on SIGVTALRM
We don't handle the SIGVTALRM signal in cpuacct_task,
so the process gets terminated, and it makes cpuacct.sh generate
a "Virtual timer expired" message for each cpuacct_task process.

In my opinion, these messages have no real value, and there is no
need in keeping them in the output.

Signed-off-by: Stanislav Kholmanskikh <stanislav.kholmanskikh@oracle.com>