syscalls/ppoll01: Fix failures with -i > 20
The raw linux ppoll syscall modifies the timespec timeout argument (in
contrast with the glibc function that hides this fact with local
variables).

Because of that the test was failing when executed with large enough -i
parameter. Each iteration would eat a bit of the timeout untill it gets
too small and we start to race with the child that sends SIGINT as the
ppoll timeouts even before the child has chance to run.

This patch fixes that by copying the timeout into a local variable
before we pass it to the ppoll syscall.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>