rtc: Display errno where applicable
RTC_ALM_READ currently fails on ppc64 4.4.21 (SLES) and the errno is not
displayed. There are not many values it is able to take, but displaying it
could still reduce the time taken to understand the problem. By the same
logic TERRNO has been added to all other test_resm calls where applicable.

Signed-off-by: Richard Palethorpe <rpalethorpe@suse.com>
Acked-by: Cyril Hrubis <chrubis@suse.cz>