test_net.sh: don't waste time waiting 1 sec in tst_ping()
The current user of this function is 'icmp-uni-basic.sh' ipsec test
group that includes 17 tests. Usually, it takes about 50 seconds for
each test run. This change allows to run the same test in ~5 seconds
because interval is 0 and default number of packets transmitted for
each size parameter increased from 10 to 500.

Signed-off-by: Alexey Kodanev <alexey.kodanev@oracle.com>
Reviewed-by: Cyril Hrubis <chrubis@suse.cz>