Use sys/xattr.h instead of attr/xattr.h
Use the glibc header sys/xattr.h instead of the libattr attr/xattr.h
header.

We were never linking againts the libattr library anyway and because of
that we picked up the glibc syscall wrappers while linking so using the
libattr header was mistake to begin with.

The only difference between the glibc header and libattr header is that
libattr defined ENOATTR alias for ENODATA errno. So in this patch we
simply replace attr/xattr.h with sys/xattr.h and ENOATTR with ENODATA.

Signed-off-by: Dejan Jovicevic <dejan.jovicevic@rt-rk.com>
Signed-off-by: Cyril Hrubis <chrubis@suse.cz>