commands: objdump01, size01: Remove.
These tests were useless since all they did was to compare command
output to itself. I.e. it called objdump or size in setup to generate
output used for comparsion in the tests later. So even if the output was
horribly wrong or empty the test would not catch anything unless the
output was actually random.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>