sched_xxx: TCONF instead of TFAIL if user is not root
Tests sched_rr_get_interval01, sched_rr_get_interval02,
sched_rr_get_interval03, sched_setparam02 and sched_setparam03
break with TFAIL if they are not ran as root. This is because
they all require setting scheduling policy to SCHED_RR or other
requiring super user privileges.
This was changed so that tests break with TCONF reporting the
appropriate message to the user if he is not root.

Signed-off-by: Dejan Jovicevic <dejan.jovicevic@rt-rk.com>
Signed-off-by: Cyril Hrubis <chrubis@suse.cz>