ipc: Update mpmsrv start/stop calls
Now that the mpmsrv is systemd based update the start/stop
calls in the test scripts. Before this patch the start/stop
calls were init.d based which fail with the systemd based
mpmsrv, and consequently also fails the test. After this patch
mpmsrv is started/stopped with sysctl (systemd based) calls

Signed-off-by: Alejandro Hernandez <ajhernandez@ti.com>
Signed-off-by: Yan Liu <yan-liu@ti.com>