commands/ld01: Fix test for LTP compiled with -m32
The test was pre-compiling the object files while LTP was build which
leads to a problem when LTP is compiled with -m32 where the produced
object files ended up incompatible with the 64bit ld command.

So after this commit the only installed datafiles are the C sources and
the object files are produced in the test setup.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>