Ensure src media files are played from local storage
When playing an audio or video file from nfs sometimes
a performance degradation have been observed. This patch
adds a donwload_to_blk_dev.sh used to download the media
into local storage to avoid the overhead of nfs. The patch
also replaces the wget calls in the gstreamer decode
and HDMI playout tests with calls to the new bash file

Signed-off-by: Alejandro Hernandez <ajhernandez@ti.com>
Signed-off-by: Yan Liu <yan-liu@ti.com>