Move gateway determination out of PRU ping LTP tests
To keep ltp tests simple because of nfs use case, the pru tests are
simplified here with dependency on host test script to set an
environment variable for gateway of pru interface. Also, adding
a setup_requires so that the tests can be run in farm setups which
have their pru interfaces connected to independent subnets.

Signed-off-by: Aparna Balasubramanian <aparnab@ti.com>
Signed-off-by: Carlos Hernandez <ceh@ti.com>