Make fps check less sensitive
This patch makes the pass/fail fps check less sensitive to small
variations. Since the test executes inside ltp-ddt which does not
appear to be a line-buffered environment there are a lot of events
that are passed to the test app which sometimes causes the fps to
fall slightly below the expected fps. This patch addresses that
problem by performing an additional check for the mean fps value
captured if the number of fps values failing is less than 5

Signed-off-by: Alejandro Hernandez <ajhernandez@ti.com>
Signed-off-by: Yan Liu <yan-liu@ti.com>