ddt/gst_dec: Run performance test on any enabled output
This patch changes gstreamer performance test to run
on any enabled output. Before this patch HDMI was not
included in the list of possible outputs and if HDMI
was the only output enabled the test would not yield
accurate results since the wrong parameters were used.
After this patch any connection enabled may be used
as an  output for the test

Signed-off-by: Alejandro Hernandez <ajhernandez@ti.com>
Signed-off-by: Yan Liu <yan-liu@ti.com>