runtest/ddt/rgx: Run func tests for finite number of frames
If number of frames (-f) is not passed, then the test cases run
forever and causes test failures due to timeouts, so pass -f
to complete tests in less time than test timeout.
Also stop weston if is running prior to calling rgx_blit_test,
otherwise the test fail because they are not expected to run inside
a window container.

Signed-off-by: Carlos Hernandez <ceh@ti.com>