testcases: ddt: scripts: alsa: Fix capture device parsing
The capture device was set to the playout device, this
worked because until now the playout and capture device
are indexed the same way. Boards like j721e* do not follow
the same convention and playout and capture devices have
different indeces. Fix the capture device parsing in the
alsa_tests.sh so that the correct capture device is used.

Signed-off-by: Alejandro Hernandez <ajhernandez@ti.com>
Signed-off-by: Yan Liu <yan-liu@ti.com>