uart_hwflow: Run test only on a particular port
In case of external loopback tests, not all instances of UART are pinned
out and not all instances are on headers where loopback can be done.
Therefore provide a way to get port per SOC and run hwflow flow control
only on that instance.

For J721e, this port is /dev/ttyS6. And hwflow tests now run only on HW
flow control.

Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Signed-off-by: Carlos Hernandez <ceh@ti.com>