ipc: Remove static marking of rpmsg rpc function index value
The function index value is currently ORed with bit 31 set to
identify it as a static registered function as per RCM Server
semantics. The rpmsg rpc driver leverages the RCM infrastructure
on the firmware side, but does not support separate static and
dynamic class of functions, and sets this bit internally. Remove
this bit to avoid any failures caused by added out-of-bound
array check in the rpmsg-rpc driver.

Signed-off-by: Suman Anna <s-anna@ti.com>