rng_test: Be more permissive
Fixes LCPD-15424

The exit value of the rngtest program is 0 only when not a single error has
been detected. However it is expected that a few errors are found when
using a large amount of data (https://jira.itg.ti.com/browse/LCPD-15424).
So instead of using the exit value to tell if the test passed or failed,
use the number of failures and a threshold.

Signed-off-by: Jean-Jacques Hiblot <jjhiblot@ti.com>