ipc: Add basic userspace rpmsg-char tests on K3 SoCs
Add a new ipc_rpmsg_char test script file that includes a bunch
of simple userspace tests using the ti_rpmsg_char library and
exercising the rpmsg_char Linux kernel driver. The tests utilize
the rpmsg_char_simple example built and packaged as part of the
ti-rpmsg-char library.

The script defines a test for each of the remote processors present
on K3 family of SoCs. Appropriate tests have to be chosen based on
the remote processors present on the SoC.

Signed-off-by: Suman Anna <s-anna@ti.com>