ddt/usbhost_eth_ping: fix usb ping test to not restore intf after test
The ddt/usbhost_eth_ping test is trying to restore and get IP address for
all existing intf at the end of the test which might fail as not all of
them may be even connected or properly configured.

The ddt/usbhost_eth_ping should manipulate only with "usb" intf and no need
to touch other intf at the end of the test as the next step is print
results and reboot.

As part of this change switch to use do_eth_up_down.sh helper to up intf.

Signed-off-by: Grygorii Strashko <grygorii.strashko@ti.com>